add controller/controller.go and split code from main.go to controller/controller.go
